# openspec/changes/dcf-ai-feedback-architecture-v1/openapi.yaml
openapi: 3.0.3
info:
  title: DCF AI Feedback Platform API
  version: 0.1.0
  description: |
    Contract baseline for event-triggered feedback lifecycle, AI classification, and supervisor triage.
    Demo mode assumptions: static logins for auth, local-first runtime, SQLite-backed persistence.
    Hugging Face config baseline (adapted from DCF-Intake demo):
    LLM_PROVIDER=huggingface, HF_MODEL=Qwen/Qwen2.5-7B-Instruct,
    HF_API_BASE=https://router.huggingface.co/v1, LLM_REQUEST_TIMEOUT_MS=600000,
    token alias support via HF_API_TOKEN or HUGGINGFACE_API_KEY.
servers:
  - url: https://api.dcf-feedback.local/v1
security:
  - demoAuthHeader: []
...
paths:
...
  /supervisor/queue:
    get:
      tags: [supervisor]
      summary: Retrieve supervisor action queue
      operationId: listSupervisorQueue
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, in_review, resolved]
        - in: query
          name: officeId
          schema:
            type: string
        - in: query
          name: userGroup
          schema:
            $ref: '#/components/schemas/UserGroup'
      responses:
        '200':
          description: Queue items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SupervisorQueueItem'
  /supervisor/queue/{queueItemId}/assign:
    post:
      tags: [supervisor]
      summary: Assign queue item
      operationId: assignQueueItem
      parameters:
        - in: path
          name: queueItemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueAssignIn'
      responses:
        '200':
          description: Queue item assigned
  /supervisor/queue/{queueItemId}/status:
    post:
      tags: [supervisor]
      summary: Update queue item status
      operationId: updateQueueStatus
      parameters:
        - in: path
          name: queueItemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueStatusIn'
      responses:
        '200':
          description: Queue item status updated
  /analytics/kpis:
    get:
      tags: [supervisor]
      summary: Retrieve KPI aggregates
      operationId: getAnalyticsKpis
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: officeId
          schema:
            type: string
        - in: query
          name: userGroup
          schema:
            $ref: '#/components/schemas/UserGroup'
      responses:
        '200':
          description: KPI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KpiResponse'
...
  /exports:
    post:
      tags: [supervisor]
      summary: Create export job
      operationId: createExportJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportJobIn'
      responses:
        '202':
          description: Export accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
  /exports/{exportId}:
    get:
      tags: [supervisor]
      summary: Get export job status
      operationId: getExportJob
      parameters:
        - in: path
          name: exportId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Export job payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJob'
  /admin/config:
    get:
      tags: [supervisor]
      summary: Retrieve admin runtime configuration
      operationId: getAdminConfig
      responses:
        '200':
          description: Current config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConfig'
    put:
      tags: [supervisor]
      summary: Update admin runtime configuration
      operationId: updateAdminConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminConfig'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConfig'
  /audit/logs:
    get:
      tags: [supervisor]
      summary: Query audit logs
      operationId: listAuditLogs
      parameters:
        - in: query
          name: actor
          schema:
            type: string
        - in: query
          name: action
          schema:
            type: string
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
...
components:
  securitySchemes:
    demoAuthHeader:
      type: apiKey
      in: header
      name: x-demo-user
      description: Static demo login identity key that maps to a predefined role profile.
...
  schemas:
    UserGroup:
      type: string
      enum: [mandated_reporter, volunteer, attorney, foster_parent]
...
    SupervisorQueueItem:
      type: object
      properties:
        queueItemId:
          type: string
        submissionId:
          type: string
        userGroup:
          $ref: '#/components/schemas/UserGroup'
        officeId:
          type: string
        status:
          type: string
          enum: [open, in_review, resolved]
        priority:
          type: string
          enum: [normal, high, critical]
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
    QueueAssignIn:
      type: object
      required: [assignee]
      properties:
        assignee:
          type: string
    QueueStatusIn:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, in_review, resolved]
        note:
          type: string
    KpiResponse:
      type: object
      properties:
        totalSubmissions:
          type: integer
        avgSentiment:
          type: number
        highUrgencyCount:
          type: integer
        responseRate:
          type: number
...
    ExportJobIn:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [csv, pdf]
        filters:
          type: object
          additionalProperties: true
    ExportJob:
      type: object
      properties:
        exportId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        format:
          type: string
          enum: [csv, pdf]
        downloadUrl:
          type: string
    AdminConfig:
      type: object
      properties:
        triggerWindows:
          type: object
          additionalProperties: true
        reminderCadence:
          type: object
          additionalProperties: true
        urgencyThreshold:
          type: number
        aiDisclosureRequired:
          type: boolean
        llmProvider:
          type: string
          enum: [huggingface]
        hfModel:
          type: string
          default: Qwen/Qwen2.5-7B-Instruct
        hfApiBase:
          type: string
          default: https://router.huggingface.co/v1
        llmRequestTimeoutMs:
          type: integer
          default: 600000
        hfTokenSource:
          type: string
          description: HF_API_TOKEN or HUGGINGFACE_API_KEY
          enum: [HF_API_TOKEN, HUGGINGFACE_API_KEY]
...
    AuditEvent:
      type: object
      properties:
        eventId:
          type: string
        actor:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          additionalProperties: true